Upgrade React and Next.js dependencies, refactor MainPageClient and ProductPage components for improved loading state handling and async data fetching. Removed unused components and updated package-lock.json for consistency.
