+++ package.json
@@ -24,12 +24,12 @@
     "classnames": "^2.5.1",
     "clsx": "^2.1.1",
     "mongodb": "^6.15.0",
-    "next": "^15.5.10",
-    "react": "18.3.1",
+    "next": "14.2.35",
+    "react": "18.2.0",
     "react-collapsible": "^2.10.0",
-    "react-dom": "18.3.1",
+    "react-dom": "18.2.0",
     "react-hook-form": "^7.54.2",
-    "react-router-dom": "^7.12.0",
+    "react-router-dom": "^7.4.0",
     "react-slick": "^0.30.3",
     "react-spring-3d-carousel": "^1.3.4",
     "react-swipeable": "^7.0.2",
