+++ src/app/[slug]/page.tsx
@@ -1,29 +1,59 @@
-import { Suspense } from "react";
-import ProductPageClient from "./ProductPageClient";
+"use client";
 
-async function ProductPageContent({
-  params
-}: {
-  params: Promise<{ slug: string }>;
-}) {
-  const { slug } = await params;
-  return <ProductPageClient slug={slug} />;
-}
+import { useEffect, useState } from "react";
+import type { Product } from "@/types/types";
+import { fetchProductById } from "@/services/products";
+import SingleProductCard from "@/components/SingleProductCard/SingleProductCard";
+import Error404 from "@/components/Error404/error404";
 
-export default function ProductPage({
-  params
-}: {
-  params: Promise<{ slug: string }>;
-}) {
-  return (
-    <Suspense
-      fallback={
-        <div className="container flex items-center justify-center min-h-[50vh]">
-          <div className="text-yellow-500 text-xl">Завантаження...</div>
-        </div>
+export default function ProductPage({ params }: { params: { slug: string } }) {
+  const [product, setProduct] = useState<Product | null>(null);
+  const [loading, setLoading] = useState(true);
+  const [error, setError] = useState(false);
+
+  useEffect(() => {
+    const loadProduct = async () => {
+      try {
+        if (!/^[a-zA-Z0-9-_]+$/.test(params.slug)) {
+          setError(true);
+          return;
+        }
+
+        const data = await fetchProductById(params.slug);
+        if (!data) {
+          setError(true);
+          return;
+        }
+        setProduct(data);
+      } catch {
+        setError(true);
+      } finally {
+        setLoading(false);
       }
-    >
-      <ProductPageContent params={params} />
-    </Suspense>
+    };
+
+    loadProduct();
+  }, [params.slug]);
+
+  if (loading) {
+    return (
+      <div className="container flex items-center justify-center min-h-[50vh]">
+        <div className="text-yellow-500 text-xl">Завантаження...</div>
+      </div>
+    );
+  }
+
+  if (error || !product) {
+    return (
+      <div className="container">
+        <Error404 />
+      </div>
+    );
+  }
+
+  return (
+    <div className="container mx-auto px-4 py-8">
+      <SingleProductCard product={product} />
+    </div>
   );
 }
