+++ src/app/page.tsx
@@ -1,5 +1,4 @@
 import type { Metadata } from "next";
-import { Suspense } from "react";
 import { MainPageClient } from "./MainPageClient";
 
 export const metadata: Metadata = {
@@ -31,9 +30,5 @@ export const metadata: Metadata = {
 };
 
 export default function MainPage() {
-  return (
-    <Suspense fallback={<div>Завантаження...</div>}>
-      <MainPageClient />
-    </Suspense>
-  );
+  return <MainPageClient />;
 }
